// frontend/src/features/projects/types.ts
export enum ProjectTaskStatus {
  Todo = 1,
  InProgress = 2,
  Done = 3,
}

export enum ProjectTaskPriority {
  Low = 1,
  Normal = 2,
  High = 3,
}
// ...
export interface ProjectTaskDto {
  id: string;
  projectId: string;
  title: string;
  description: string | null;
  status: ProjectTaskStatus;
  priority: ProjectTaskPriority;
  dueDate: string | null;
  assignedUserId: string | null;
  createdByUserId?: string | null;
  createdAt: string;
  updatedAt: string;
  concurrencyStamp: string;
  labels: string[];
}